Update CPL to CPL_1.10.3_300618
Changelog (1.10.3):

The height of the dock of the bar is fixed on any device with any settings The 
number of lines in the menu should match the selected number 
Fixed a bug where icons on the desktop were not 'pressed' The 
clock in actions with the 'Countdown' widget starts any, not just Google Clock 
Improved layers of the icon editor and settings 
For devices without the on-screen buttons added setting 'Additional indent of the search string' - now the position of the icons in the dock should be centered. 
For devices 6+ the main menu of settings in the style of Android P 
Possibility ь hide the search bar in the application menu 
Ability to select buttons in the 'classic' search bar (paid functionality)
The text in the dock bar, which can be configured (included with the microphone in the docking bar). The 
ability to change the text color in Google search in the application menu (7+) 
Part of the graphics updated in the style of Android P 
Option to display widgets in full size fixed (I have to choose between the desktop or widgets in full size, I added a note to the setting) 
New form of adaptive Cylinder icons (8+) 
Other minor fixes